Add files via upload
This repository features a deep learning project designed to classify plant diseases from leaf images using the PlantVillage dataset. The workflow involves loading and preprocessing the dataset, building and training an autoencoder for image reconstruction, and leveraging the encoder part of the autoencoder to construct a classifier for disease prediction. The project starts by resizing and normalizing the images, then splits the data into training and validation sets.

The autoencoder, comprising an encoder and a decoder, is trained to reconstruct images, capturing essential features in a compressed form. The encoder’s output is then used in a classifier, which adds dense layers for predicting the disease class of the input images. After training, the classifier’s performance is evaluated on the validation dataset, and its accuracy is printed. To visualize the autoencoder’s effectiveness, original and reconstructed images are displayed side by side.

Moreover, the project includes a detailed analysis of misclassified cases by comparing the original and reconstructed images. This step helps identify why certain predictions are incorrect, shedding light on potential areas for model improvement. The provided functions also predict and display true and false cases, offering insights into the model’s strengths and weaknesses. Overall, this repository offers a robust approach to plant disease classification, combining image reconstruction and classification to enhance predictive accuracy and understand model limitations.